backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1
setuptools>=65.5.1
django-storages==1.14.2
boto3==1.34.34
